Complete interface testing and stubbing, param validate WIP [ci skip]
