+++ gems/aws-sdk-core/lib/aws-sdk-core/stubbing/protocols/rest.rb
@@ -1,3 +1,5 @@
+require 'aws-eventstream'
+
 module Aws
   module Stubbing
     module Protocols
@@ -72,6 +74,46 @@ module Aws
 
         def head_operation(operation)
           operation.http_method == "HEAD"
+        end
+
+        def eventstream?(rules)
+          rules.shape.members.each do |_, ref|
+            return true if ref.eventstream
+          end
+          false
+        end
+
+        def encode_eventstream_response(rules, data, builder)
+          data.inject('') do |stream, event_data|
+            # construct message headers and payload
+            opts = {headers: {}}
+            event_ref = rules.shape.member(event_data.delete(:event_type))
+            event_data.each do |k, v|
+              member_ref = event_ref.shape.member(k)
+              if member_ref.eventheader
+                opts[:headers][member_ref.location_name] = Aws::EventStream::HeaderValue.new(
+                  value: v,
+                  type: member_ref.eventheader_type
+                )
+              elsif member_ref.eventpayload
+                case member_ref.eventpayload_type
+                when 'string'
+                  opts[:payload] = StringIO.new(v)
+                when 'blob'
+                  opts[:payload] = v
+                when 'structure'
+                  opts[:payload] = StringIO.new(builder.new(member_ref).serialize(v))
+                end
+              end
+            end
+            opts[:headers][':event-type'] = Aws::EventStream::HeaderValue.new(
+              value: event_ref.location_name,
+              type: "string"
+            )
+            stream << Aws::EventStream::Encoder.new.encode(
+              Aws::EventStream::Message.new(opts))
+            stream
+          end
         end
 
       end
