+++ gems/aws-sdk-core/lib/aws-sdk-core/stubbing/protocols/rest_xml.rb
@@ -1,5 +1,3 @@
-require 'aws-eventstream'
-
 module Aws
   module Stubbing
     module Protocols
@@ -9,9 +7,7 @@ module Aws
 
         def body_for(api, operation, rules, data)
           if eventstream?(rules)
-            # TODO Extract this
-            encode_response(rules, data)
-            # stub_eventstream # from REST
+            encode_eventstream_response(rules, data, Xml::Builder)
           else
             xml = []
             rules.location_name = operation.name + 'Result'
@@ -30,47 +26,6 @@ module Aws
 
         def xmlns(api)
           api.metadata['xmlNamespace']
-        end
-
-        private
-
-        def eventstream?(rules)
-          rules.shape.members.each do |_, ref|
-            return true if ref.eventstream
-          end
-        end
-
-        def encode_response(rules, data)
-          data.inject('') do |stream, event_data|
-            # construct message headers and payload
-            opts = {headers: {}}
-            event_ref = rules.shape.member(event_data.delete(:event_type))
-            event_data.each do |k, v|
-              member_ref = event_ref.shape.member(k)
-              if member_ref.eventheader
-                opts[:headers][member_ref.location_name] = Aws::EventStream::HeaderValue.new(
-                  value: v,
-                  type: member_ref.eventheader_type
-                )
-              elsif member_ref.eventpayload
-                case member_ref.eventpayload_type
-                when 'string'
-                  opts[:payload] = StringIO.new(v)
-                when 'blob'
-                  opts[:payload] = v
-                when 'structure'
-                  #TODO
-                end
-              end
-            end
-            opts[:headers][':event-type'] = Aws::EventStream::HeaderValue.new(
-              value: event_ref.location_name,
-              type: "string"
-            )
-            stream << Aws::EventStream::Encoder.new.encode(
-              Aws::EventStream::Message.new(opts))
-            stream
-          end
         end
 
       end
